Changed Thread Sizes and Wheel Speed Integer Size
-Wheel Speed is calculating absolute velocity, but not direction

+++ teensy_chibios/src/main/include/system_data.h
@@ -6,7 +6,7 @@
 typedef struct sensor_data_t {
    int16_t imu_angle;
    int16_t tof_dist;
-   int16_t wheel_speed;
+   int32_t wheel_speed;
    int16_t right_URF;
    int16_t left_URF;
    int16_t rear_URF;
